The info dialog from the options menu in the data activities should display a different message depending on what tab is currently visible.  The messages should be specific to whether the data being displayed is for all apps or just one.
