+++ app/src/main/java/com/PrivacyGuard/Application/Activities/AppDataActivity.java
@@ -2,6 +2,8 @@ package com.PrivacyGuard.Application.Activities;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.support.v7.app.AlertDialog;
+import android.view.MenuItem;
 
 import com.PrivacyGuard.Application.Database.DatabaseHandler;
 import com.PrivacyGuard.Plugin.LeakReport;
@@ -41,5 +43,49 @@ public class AppDataActivity extends DataActivity {
     @Override
     public String getAppPackageName() {
         return packageName;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch(item.getItemId()) {
+            case R.id.info:
+                AlertDialog alertDialog;
+
+                switch (tabLayout.getSelectedTabPosition()) {
+                    case 0:
+                        alertDialog = new AlertDialog.Builder(this)
+                                .setTitle(R.string.leak_report_title)
+                                .setIcon(R.drawable.info_outline)
+                                .setMessage(R.string.report_message_single_app)
+                                .setPositiveButton(R.string.dialog_accept, null)
+                                .create();
+                        alertDialog.show();
+                        return true;
+
+                    case 1:
+                        alertDialog = new AlertDialog.Builder(this)
+                                .setTitle(R.string.leak_summary_title)
+                                .setIcon(R.drawable.info_outline)
+                                .setMessage(R.string.summary_message_single_app)
+                                .setPositiveButton(R.string.dialog_accept, null)
+                                .create();
+                        alertDialog.show();
+                        return true;
+
+                    case 2:
+                        alertDialog = new AlertDialog.Builder(this)
+                                .setTitle(R.string.leak_query_title)
+                                .setIcon(R.drawable.info_outline)
+                                .setMessage(R.string.query_message_single_app)
+                                .setPositiveButton(R.string.dialog_accept, null)
+                                .create();
+                        alertDialog.show();
+                        return true;
+                }
+
+                return false;
+        }
+
+        return super.onOptionsItemSelected(item);
     }
 }
