+++ app/src/main/res/values/strings.xml
@@ -64,13 +64,21 @@
     <string name="keyword_category">Keyword</string>
 
     <!-- Android Plot Display -->
-    <string name="graph_message">This feature allows you to see when your apps are leaking compared to when the app is in the foreground or background.</string>
     <string name="dialog_accept">OK</string>
-    <string name="leak_report_title">Leak Report</string>
     <string name="app_status_foreground">Foreground</string>
     <string name="app_status_background">Background</string>
     <string name="app_status_no_data">No Data</string>
     <string name="app_data_info">Info</string>
+    <string name="leak_report_title">Leak Report</string>
+    <string name="leak_summary_title">Leak Summary</string>
+    <string name="leak_query_title">Leak Query</string>
+    <string name="report_message_single_app">This graph shows when your app is leaking compared to when it is in the foreground or background. Use the navigation arrows to move forward and backward in time.</string>
+    <string name="summary_message_single_app">A brief summary of the leaks for your app.</string>
+    <string name="query_message_single_app">Set the parameters as you wish in order to investigate what your app has been leaking.</string>
+    <string name="report_message_all_apps">This graph shows when your apps are leaking. Use the navigation arrows to move forward and backward in time.</string>
+    <string name="summary_message_all_apps">A brief summary of the leaks for all your apps.</string>
+    <string name="query_message_all_apps">Set the parameters as you wish in order to investigate what your apps have been leaking.</string>
+
 
     <!-- Spinners -->
     <string name="spinner_category">Category:</string>
